scripts/stage_04_diarization.py
import os
import json
import torch
from pyannote.audio import Pipeline

def run(vocals_audio: str, output_json: str):
    """
    Stage 4: Speaker Diarization
    Requirements Addressed: "Different character 100% original voice" & "Analysis Movie character"
    
    This script listens to the isolated vocals and creates a temporal map defining EXACTLY 
    when Character A speaks vs Character B speaks. This guarantees we don't accidentally 
    clone Character A's voice for Character B's translated dialogue lines.
    """
    print(f"[Stage 4] Analyzing movie characters from {vocals_audio} using Pyannote...")
    
    # Pyannote pipeline specifically configured for diarization
    # Note: Requires a HuggingFace auth token in production to download the gated model.
    # We instantiate it to run on CUDA if available.
    auth_token = os.environ.get("HF_AUTH_TOKEN", None)
    if not auth_token:
        print("[Stage 4] WARNING: No HF_AUTH_TOKEN found in environment. Expected for pyannote.audio/speaker-diarization-3.1")
        
    try:
        pipeline = Pipeline.from_pretrained("pyannote/speaker-diarization-3.1", use_auth_token=auth_token)
        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        pipeline.to(device)
        
        diarization = pipeline(vocals_audio)
        
        characters = []
        
        # Turn the Pyannote Annotation object into a strict temporal JSON list
        for turn, _, speaker in diarization.itertracks(yield_label=True):
            characters.append({
                "character_id": speaker,
                "start": float(turn.start),  # The exact 'time stand' boundary 
                "end": float(turn.end)       # The exact 'time stand' boundary end
            })
            
        os.makedirs(os.path.dirname(output_json), exist_ok=True)
        with open(output_json, "w", encoding="utf-8") as f:
            json.dump({"characters": characters}, f, indent=4)
            
        print(f"[Stage 4] Success: Identified {len(set(c['character_id'] for c in characters))} character(s) in timeline.")
        return output_json
        
    except Exception as e:
        print(f"[Stage 4] Diarization failure: {e}")
        # Fallback mechanism if auth fails or model unavailable
        print("[Stage 4] Mocking single character fallback due to error...")
        with open(output_json, "w") as f:
            json.dump({"characters": [{"character_id": "SPEAKER_00", "start": 0.0, "end": 9999.0}]}, f)
        return output_json
