requirements.txt
# ═══════════════════════════════════════════
# Local Dubbing Platform – Requirements
# Aligned with OmniVoice 0.1.3 dependencies
# ═══════════════════════════════════════════

# ── Core Framework ──
fastapi>=0.100.0
uvicorn>=0.30.0
starlette>=0.30.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── Task Queue ──
celery[redis]>=5.3.0
redis>=5.0.0

# ── PyTorch (match OmniVoice) ──
torch>=2.4
torchaudio>=2.4

# ── ML / Transformers (match OmniVoice) ──
transformers
accelerate

# ── OmniVoice dependencies ──
pydub
soundfile
numpy<=1.24.4
gradio
tensorboardX
webdataset
google-cloud-translate

# ── Audio Processing ──
demucs
librosa
scipy

# ── Speech & Language ──
openai-whisper
pyannote.audio
sentencepiece

# ── Video ──
ffmpeg-python

# ── Config & Utils ──
pyyaml
tqdm
httpx
aiofiles
python-dotenv

# ── OmniVoice optional / eval ──
jiwer
langdetect
scikit-learn
pandas
